VCV-150: Refactor folder structure for operations feature
