+++ src/app/(dashboard)/page.tsx
@@ -1,4 +1,4 @@
-import DashboardPageClient from '@/features/dashboard/page-client/dashboard-page-client';
+import DashboardPageClient from '@/features/dashboard/components/dashboard-page-client';
 
 export default function DashboardPage() {
     return <DashboardPageClient />;
