+++ src/features/operations/ai-performance/components/ai-performance-summary-card.tsx
