+++ src/features/operations/field-user-compliance/components/operations-field-user-compliance.tsx
@@ -3,10 +3,10 @@
 import { useGetAllSessions } from '@/api/session/hooks/use-get-all-sessions';
 import { Skeleton } from '@/components/ui/skeleton';
 import type { LocationQueryParam } from '@/lib/location/location-query';
-import { buildFieldUserComplianceData } from '@/features/operations/utils/field-user-compliance-data';
+import { buildFieldUserComplianceData } from '@/features/operations/field-user-compliance/utils/field-user-compliance-data';
 import { eachMonthOfInterval, format, parseISO } from 'date-fns';
 import { useMemo } from 'react';
-import FieldUserComplianceChart from '@/features/operations/components/field-user-compliance/field-user-compliance-chart';
+import FieldUserComplianceChart from '@/features/operations/field-user-compliance/components/field-user-compliance-chart';
 
 interface OperationsFieldUserComplianceProps {
     locationQueryParam: LocationQueryParam;
