+++ src/features/operations/field-user-compliance/utils/build-session-chart-pills.ts
