+++ src/features/operations/geographical-summary/components/map-navigator.tsx
