+++ src/features/operations/geographical-summary/hooks/use-site-geocode.ts
@@ -4,7 +4,7 @@ import { geocodeKeys } from '@/api/geocode/geocode-keys';
 import { fetchGeocode } from '@/api/geocode/hooks/use-get-geocode';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { useQueryClient } from '@tanstack/react-query';
-import type { SiteMarker } from '@/features/operations/utils/site-marker-data';
+import type { SiteMarker } from '@/features/operations/geographical-summary/utils/site-marker-data';
 import type { Geocode } from '@/api/geocode/validation/geocode-schema';
 
 export function useSiteGeocode(markers: SiteMarker[]): {
