+++ src/features/operations/geographical-summary/utils/site-marker-data.ts
